업데이트된 like_cnt 데이터베이스에 저장

+++ src/main/java/com/team2/_3dinterest/domain/ahyeon/heart/Heart.java
@@ -6,7 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Heart {
-    private Long heartId;
-    private Long postId;
+    private int heartId;
+    private int postId;
     private Long userId;
+
 }
