+++ src/main/java/com/team2/_3dinterest/domain/ahyeon/post/service/PostService.java
@@ -6,8 +6,8 @@ import com.team2._3dinterest.domain.ahyeon.post.Post;
 
 
 public interface PostService {
-    Post getPostById(Long postId);
-    Post updateLikeAndReturnPost(Long postId, String userId);
+    Post getPostById(int postId);
+    Post updateLikeAndReturnPost(int postId, String userId);
 
 
 }
